handle ties for most-probable soil
Ties between most-probable soils were being mapped out most-least
probable in (essentially) alphabetical order. Now they're shuffled so
ties come out in a random sequence, This removes the prediction bias but
does mean that the outputs of this function will vary slightly if
repeated without a seed set. Thanks to Jon Walton (QLD DNRM) for asking
the question that led to this amendment.